Remove graph parameter from CSM __init__
This was mainly used to create copies in the
time interpolation functions and shouldn't be exposed
to the public in the first place. Howeve, touching
private members of the copy during interpolation
would violate some best practice rules and static
analyzers wuld complain. This is now solved due to
the possibility to use in place interpolations.