Schema updates for current and upcoming asdf changes (#910)
* updates schema ids to match uris

* update uri prefxi in update_manifest

* restrict schema tests using ndarray-1.0.0 to stardard 1.5.0

The next version of the ASDF standard (1.6.0) contains a new
version of the ndarray schema (ndarray-1.1.0). Examples that
explicitly use ndarray-1.0.0 will only work with the older
standard versions (<1.6.0). This doesn't impact file
reading and writing (as if 1.5.0 is used for write, ndarray-1.0.0
will be used, if 1.6.0 is used, ndarray-1.1.0 will be used). However
the schema examples require hard-coded yaml that is less flexible
than real files and require a restriction like that added in
this commit.

* take 2 at trying to fix update_manifest script

* ruff ruff :wag:

* import SerializationContext from asdf.extension for version >3

* add changelog

* remove asdf pin

* update changelog

* remove support for AsdfSpec in format_tag

* validate tree provided to WeldxFile

* Update file.py doctest

* add asdf version to test matrix

* update artifact condition

---------

Co-authored-by: Çağtay Fabry <cagtay.fabry@bam.de>